crate-universe allows workspaces with MODULE.bazel only (#2720)
According to the [migration
guide](https://bazel.build/external/migration), `WORKSPACE.bazel` is no
longer required with Bazel 6.3 or later. This updates `crate_universe`
to respect that, allowing `MODULE.bazel` to be the root marker of a
workspace, without requiring a `WORKSPACE` or `WORKSPACE.bazel`.

Happy to iterate on this.