Allow for unnamed workspaces/modules when using crates_vendor (#2730)
Addresses https://github.com/bazelbuild/rules_rust/issues/2661

If the workspace name is empty, the crate_module_template will use a
label like `Label(//<stuff>)` instead of `Label(@//<stuff>)`

Most of the changes in this are just regenerating the vendor'ed crates
in the examples. I ran the tests in all the relevant examples to ensure
they still worked.