crate_universe: Only compile label regex once. (#2751)
I hit this when profiling for something else, we spent 1.1sec of 1.3sec
of `cargo-bazel generate` compiling this regular expression. It seems to
be quite bad on the regex compiler, as even compiling it once takes 8ms.

I had to add the `once_cell` crate to unwrap the Result type - the
corresponding stdlib function is still nightly-only unfortunately.
However, that crate was already in the transitive dependency tree, so
it's a net-same change.