Updated Bazelmod cross compilation to catch MacOS sys dependency issues (#2926)
As discussed in issue
https://github.com/bazelbuild/rules_rust/issues/2917,
this updated cross compilation example serves as a bridge until the
exact cause of the sys crates not building on Mac issue has been found
and better regression testing has been added.

Also, I took the freedom to trim and streamline the LLVM setup making it
easier to configure while preserving the property that it caches failing
sys crate builds. Updated the Readme accordingly.

---------

Signed-off-by: Marvin Hansen <marvin.hansen@gmail.com>