Add per-target link flags in the right place (#2963)
Previously we were adding these early on in the link args.

This was a bug - if some transitive dep foo has a linkopt of `-lbar`,
the `-lbar` needs to be later in the link args list than `-lfoo`.

Rather than arbitrarily adding them early on and hoping things work,
this adds the args for each library just after we add a dependency on
that library.