Make toolchain vars available but don't automatically set them (#2969)
Previously if a target depended on a toolchain which exposed `$(FOO)`,
we would automatically set `$FOO` as an env var.

This was not correct. Toolchains should make Make variables available
for expansion in env vars, but should not automatically promote them to
action env.

If you want `$FOO` set, you should set `$FOO=$(FOO)` in the appropriate
env var attribute.