Make lto toolchain options public and don't link proc-macros with LTO (#3147)
I would like to make lto setting public in toolchain so that when we are
providing custom toolchains it would be easier to make configurations
for LTO with `select` statements.

The logic which determines the flags based on cargo outputs.

Proc-macro crates can not be linked with lto and we should not emit
bitcode either. This fixes #3143

---------

Co-authored-by: Daniel Wagner-Hall <dwagnerhall@apple.com>