feat: Add `rust_lint_group` and `cargo_lints` rule to define lint groups (#2993)
Chatted about this previously in
[Slack](https://bazelbuild.slack.com/archives/CSV56UT0F/p1718030356910019).
The goal with this PR is to make it easier to define and share lint
configurations for your build.

### Description

The PR adds two new rules, `rust_lint_group` and `cargo_lints`. It also
adds a "lints" attr to `rust_library` and `rust_binary` rules that
expects a new `LintsInfo` provider.

* `rust_lint_group` allows you to define Rust, Clippy, and Rustdoc lints
in a `BUILD` file.
* `cargo_lints` automatically generates a set of lints by reading a
crate's `Cargo.toml`, and optionally the workspace's Cargo.toml for
workspace inheritance.

Then the rustc, clippy, and rustdoc actions are all updated to check for
the `LintsInfo` provider and appends the correct arguments so the lints
take effect.


### Design

Honestly this is my first large change to a set of Bazel rules, so I
definitely could have done something wrong here!

The change is split into two commits, the first introduces
`rust_lint_group` which IMO is relatively straight forward. Some
attributes are defined on a rule which are then formatted into command
line flags and passed around with a provider.

The second commit adds `cargo_lints` and is much larger, a few things
worth considering:

* I was back and forth a bit on using a `repository_rule` or a regular
`rule`. While the `repository_rule` maps well to a Cargo Workspace and
you'd only need to define it once, not everyone uses workspaces and so I
figured a regular `rule` was more general.
* Parsing a `Cargo.toml` is done via a new Rust binary called
`cargo_toml_info`. I tried to keep the external dependencies on this
binary to a minimum, it only has one at the moment which is
[`cargo_toml`](https://docs.rs/cargo_toml/latest/cargo_toml/) and I
based this change largely off of
https://github.com/bazelbuild/rules_rust/pull/2772. I tried to make the
tool general though so other Cargo metadata rules could use it in the
future.

### Tests

There aren't any! I wasn't sure where the best place to start was, any
guidance here is appreciated!