Reduce number of default `supported_platform_triples` for crate_universe (#3239)
This should save substantial time for users of `crate_universe` as they
don't need to do splicing for platforms they're not going to use. It's
still recommended users explicitly set `supported_platform_triples` but
my expectation with this change is that it covers most of the bases for
users to still provide them with a sufficiently responsive experience
compared to the 34 platforms that used to be in splicing.