Use `--remote_download_all` when building `rust_analyzer_crate_spec`s (#3192)
This fixes #3170, by ensuring that transitive deps like procmacro dylibs
are fetched from the disk-cache.

This works in my case, but it's unclear to me if this has unintended
consequences for people running remote builds etc, and this can easily
be addressed by users manually by adding `build --remote_download_all`
to their .bazelrc too if you prefer not merging this.