Fixed loss of data in gen_rust_project for duplicate crate ids (#3317)
Previously `aliases` and `source` were susceptible to being dropped from
`rust-project.json` in the event a crate without either of these fields
was processed before another crate with them. These are now aggregated
and `gen_rust_project` should be more resilient to the ordering targets
are listed in aquery outputs.