cargo_toml_env_vars: rule for generating env vars from a Cargo.toml file (#2772)
This allows people to set all of the compile-time variables that would
be set by Cargo.

This doesn't currently wire up crate_universe to generate and depend on
these targets for third-party crates, but I can add that if this PR
looks good - it should just be a matter of a small addition to the
generation templates (one new target, and setting `rustc_env_files` to
point at the new target).

Work towards #2677