fix: Duplicate aliases generated in rendering when two crates of the same version (#3386)
Original cause:
https://github.com/bazelbuild/rules_rust/commit/5e904830eee39b37775855c6ff7e7e1079eb59e0#diff-52bdc04b030a02035a19a3c805288b7720191185a18f6b8f16f61a59baa42d3fR224
Related to: https://github.com/bazelbuild/rules_rust/pull/3215

This PR updates rendering in `cargo-bazel` to handle the case when there
are multiple dependencies on a crate at the same version, but they're
aliased. For example, when a `Cargo.toml` has the following:

```
itertools = "0.11.24"
itertools_other = { version = "0.11.24", package = "itertools" }
```

Previously this would generate two aliases both with the name
`itertools-0.11.24` but now we only create a single alias of that name.

Encountering this case is odd but not uncommon, especially if you also
use a tool like
[`cargo-hakari`](https://github.com/guppy-rs/guppy/tree/main/tools/cargo-hakari)
in your workspace.