[Doc] Improve documentation on toolchain registration with bzlmod (#3430)
Hi, this is referencing issue #2973 , i'm sorry i kinda forgot this was
open too. Let me know if it is clear or if it needs modification.

Parallelly, i removed refernces to a no longer existing file
`flatten.md`. Is this correct, or do you want to open a separate
issue/pr? Thanks!